+++ libs/ledger-live-common/src/bot/formatters.ts
@@ -8,9 +8,20 @@ import {
 import { formatCurrencyUnit } from "../currencies";
 import type { MutationReport, AppCandidate } from "./types";
 import type { Transaction } from "../generated/types";
-export const formatTime = (t: number) =>
+
+const formatTimeMinSec = (t: number) => {
+  const totalsecs = Math.round(t / 1000);
+  const min = Math.floor(totalsecs / 60);
+  const sec = totalsecs - min * 60;
+  if (!sec) return `${min}min`;
+  return `${min}min ${sec}s`;
+};
+
+export const formatTime = (t: number): string =>
   t > 3000
-    ? `${Math.round(t / 100) / 10}s`
+    ? t > 100000
+      ? formatTimeMinSec(t)
+      : `${Math.round(t / 100) / 10}s`
     : `${t < 5 ? t.toFixed(2) : t.toFixed(0)}ms`;
 
 const formatDt = (from, to) => (from && to ? formatTime(to - from) : "?");
