--- libs/ledger-live-common/src/bot/test.bot.ts
@@ -1,5 +1,0 @@
-import "../__tests__/test-helpers/setup";
-import { run } from "./run";
-
-jest.setTimeout(110 * 60 * 1000);
-test("bot", run);
