+++ libs/ledger-live-common/src/env.ts
@@ -520,7 +520,7 @@ const envDefinitions = {
     desc: "maximum limit to synchronize accounts concurrently to limit overload",
   },
   BOT_MAX_CONCURRENT: {
-    def: 4,
+    def: 5,
     parser: intParser,
     desc: "maximum limit to run bot spec in parallel",
   },
